fix password regex, extract input props, add errormessages
