+++ src/Components/EntryForm/EntryForm.stories.tsx
@@ -7,12 +7,18 @@ export default {
   component: EntryForm,
 } as ComponentMeta<typeof EntryForm>;
 
-const noop = () => {};
-
 const Template: ComponentStory<typeof EntryForm> = (args) => <EntryForm {...args} />;
 
 export const DynamicForm = Template.bind({});
 DynamicForm.args = {
-  loginFn: noop,
-  registerFn: noop,
+  loginFn: (form: HTMLFormElement) => {
+    console.log(`Email: ${(form.elements.namedItem('login-email') as HTMLInputElement).value}`);
+    console.log(`Password: ${(form.elements.namedItem('login-password') as HTMLInputElement).value}`);
+  },
+  registerFn: (form: HTMLFormElement) => {
+    console.log(`Name: ${(form.elements.namedItem('register-name') as HTMLInputElement).value}`);
+    console.log(`User: ${(form.elements.namedItem('register-user') as HTMLInputElement).value}`);
+    console.log(`Email: ${(form.elements.namedItem('register-email') as HTMLInputElement).value}`);
+    console.log(`Password: ${(form.elements.namedItem('register-password') as HTMLInputElement).value}`);
+  },
 };
