+++ src/Components/Input/Input.tsx
@@ -2,20 +2,7 @@ import React, { FC, useEffect, useRef, useState } from 'react';
 import { v4 as uuid } from 'uuid';
 import { Icon, IconColor } from '../Icon/Icon';
 import { defaultErrorMessages, getTranslationKeyForError } from './input-validation.helpers';
-
-export type InputType = 'text' | 'password' | 'email' | 'textarea';
-export interface InputProps {
-  label: string;
-  name: string;
-  type: InputType;
-  required: boolean;
-  validationTrigger: number;
-  errorTranslations?: Record<string, string>;
-  minLength?: number;
-  maxLength?: number;
-  pattern?: string;
-  placeholder?: string;
-}
+import { InputProps, InputType } from '../../models';
 
 interface InputState {
   value: string;
