+++ src/models/models.ts
@@ -1,1 +1,2 @@
 export type ButtonType = 'button' | 'submit' | 'reset';
+export type InputType = 'text' | 'password' | 'email' | 'textarea';
