+++ src/models/props.ts
@@ -1,4 +1,4 @@
-import { ButtonType } from './models';
+import { ButtonType, InputType } from './models';
 
 export interface ButtonProps {
   id: string;
@@ -26,3 +26,33 @@ export interface ButtonGroupProps {
   RIcon?: string;
   ROnClick?: () => void;
 }
+
+export interface FormProps {
+  isRegistered: boolean;
+  errorTranslations?: Record<string, string>;
+  loginFn: (form: HTMLFormElement) => void;
+  registerFn: (form: HTMLFormElement) => void;
+}
+
+export interface InputProps {
+  label: string;
+  name: string;
+  type: InputType;
+  required: boolean;
+  validationTrigger: number;
+  errorTranslations?: Record<string, string>;
+  minLength?: number;
+  maxLength?: number;
+  pattern?: string;
+  placeholder?: string;
+}
+
+export interface FileInputProps {
+    title: string;
+    restrictions: string;
+    label: string;
+    errorTranslations?: Record<string, string>;
+    allowedFileSize?: number;
+    allowedExtensions?: RegExp;
+    addFileFn: (file: File) => void;
+  }
